Introduced Schema flags

+++ Library/src/uk/ac/ucl/excites/sapelli/transmission/TransmissionClient.java
@@ -45,6 +45,23 @@ public abstract class TransmissionClient extends StorageClient
 
 	// STATICS-------------------------------------------------------
 	static public final long TRANSMISSION_MANAGEMENT_MODEL_ID = 0; // reserved!
+	
+	/**
+	 * Flag indicating that a Schema has been defined at the Transmission layer of the Sapelli Library
+	 */
+	static private final int SCHEMA_FLAG_TRANSMISSION_LAYER =	1 << 6;
+	
+	/**
+	 * Schema flag indicating that records of the Schema can be transmitted using the Transmission/Payload classes
+	 */
+	static public final int SCHEMA_FLAG_TRANSMITTABLE = 		1 << 7;
+	
+	// Note: flag bits 8 & 9 are reserved for future Transmission layer usage
+	
+	/**
+	 * Flags used on "internal" Transmission layer Schemata
+	 */
+	static public final int SCHEMA_FLAGS_TRANSMISSION_INTERNAL = SCHEMA_FLAG_TRANSMISSION_LAYER;
 
 	// DYNAMICS------------------------------------------------------
 	public final StoreHandle<TransmissionStore> transmissionStoreHandle = new StoreHandle<TransmissionStore>(new StoreCreator<TransmissionStore>()
@@ -157,23 +174,22 @@ public abstract class TransmissionClient extends StorageClient
 		@Override
 		public void storageEvent(RecordOperation operation, RecordReference recordRef)
 		{
-			// TODO if(/*check if transmittable*/true)
-			
-			receiverLoop : for(Correspondent receiver : getReceiversFor(recordRef.getReferencedSchema()))
-			{
-				switch(operation)
+			if(recordRef.getReferencedSchema().hasFlags(SCHEMA_FLAG_TRANSMITTABLE))
+				receiverLoop : for(Correspondent receiver : getReceiversFor(recordRef.getReferencedSchema()))
 				{
-					case Inserted :
-					case Updated :
-						tStore.storeTransmittableRecord(receiver, recordRef, null); // will this wipe tosend rec's for same rec that already had a transmission?
-						break;
-					case Deleted :
-						tStore.deleteTransmittableRecord(recordRef); // record will be forgotten about for each receiver ...
-						break receiverLoop; // ... so we are done here
-					default :
-						throw new IllegalArgumentException("Unknown " + RecordOperation.class.getSimpleName());
+					switch(operation)
+					{
+						case Inserted :
+						case Updated :
+							tStore.storeTransmittableRecord(receiver, recordRef, null); //TODO will this wipe tosend rec's for same rec that already had a transmission?
+							break;
+						case Deleted :
+							tStore.deleteTransmittableRecord(recordRef); // record will be forgotten about for each receiver ...
+							break receiverLoop; // ... so we are done here
+						default :
+							throw new IllegalArgumentException("Unknown " + RecordOperation.class.getSimpleName());
+					}
 				}
-			}
 		}
 		
 		@Override
